% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_als.r
\name{alm.ALS}
\alias{alm.ALS}
\title{run function als}
\usage{
alm.ALS(Y, X, tol = 1e-05, imax = 100, verbose = FALSE, seed = 1)
}
\arguments{
\item{Y}{some variable X some variable}
}
\value{
a list
}
\description{
More details
}
