% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfunctions.r
\name{amprod}
\alias{amprod}
\title{Array-matrix product}
\usage{
amprod(A, M, k)
}
\arguments{
\item{A}{a real valued array}

\item{M}{a real matrix}

\item{k}{an integer, a mode of \code{A}}
}
\description{
Multiply an array by a matrix along a given mode
}
\details{
This function multiplies a matricized array by another
matrix, and then reforms the result into a new array.
}
\examples{
A<-rsan(c(5,4,3))
B<-rsan(c(2,5))
amprod(A,B,1)
}
\author{
Peter Hoff
}
\keyword{arrays}
