% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\name{gfriends}
\alias{gfriends}
\title{Girls high school friendship data}
\format{A list including the following dyadic and nodal variables:
\itemize{
\item \code{Y}: an undirected sociomatrix of friendship ties;
\item \code{X}: nodal covariates including grade and normalized gpa and smoking scores.
}}
\description{
Directed inditicator of friendship between 144 high school girls
}
