% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\name{htmanagers}
\alias{htmanagers}
\title{Krackhardt's  high-tech managers}
\format{A list consisting of the following:
\itemize{
\item \code{Y}: a directed socioarray representing to whom the ego
 goes to for advice, whom the ego considers a friend, and to whom the
 ego reports to.
\item \code{X}: nodal covariates including age, length of service, level in the corporate hierarchy and department.
}}
\source{
\url{http://moreno.ss.uci.edu/data.html#krackht}
}
\description{
Self-reported binary relations among 21 managers of a high-tech
company, as well as several nodal attributes.
}
