% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfunctions.r
\name{kron}
\alias{kron}
\title{Kronecker product}
\usage{
kron(...)

kron(...)
}
\arguments{
\item{\ldots}{Either a sequence of matrices, or a single list object of matrices}

\item{\ldots}{Either a sequence of matrices, or a single list object of matrices}
}
\description{
Calculate a kronecker product

Calculate a kronecker product
}
\details{
This function calculates the Kronecker product of a list
of matrices.

This function calculates the Kronecker product of a list
of matrices.
}
\examples{
A<-rsan(c(2,3)) ; B<-rsan(c(3,4)) ; C<-rsan(c(4,5))
kron(A,B,C)
M<-list(A,B,C)
kron(M)
A<-rsan(c(2,3)) ; B<-rsan(c(3,4)) ; C<-rsan(c(4,5))
kron(A,B,C)
M<-list(A,B,C)
kron(M)
}
\author{
Peter Hoff

Peter Hoff
}
\keyword{matrices}
