% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfunctions.r
\name{ksum}
\alias{ksum}
\title{Kronecker sum}
\usage{
ksum(...)
}
\arguments{
\item{\ldots}{Either a sequence of matrices, or a single list object of matrices}
}
\description{
Calculate a kronecker product
}
\details{
This function calculates the Kronecker sum of a list
of matrices.
}
\examples{
A<-rsan(c(2,3)) ; B<-rsan(c(3,4)) ; C<-rsan(c(4,5))
ksum(A,B,C)
M<-list(A,B,C)
ksum(M)
}
\author{
Peter Hoff
}
\keyword{matrices}
