% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfunctions.r
\name{mat}
\alias{mat}
\title{Matricization}
\usage{
mat(A, k)
}
\arguments{
\item{A}{an array}

\item{k}{a mode of \code{A}, along which to matricize}
}
\description{
Matricize an array
}
\details{
This functions matricizes an array along a given mode.
}
\examples{
A<-rsan(4,3,2)
mat(A,2)
}
\keyword{arrays}
\keyword{matrices}
