% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfunctions.r
\name{mhalf}
\alias{mhalf}
\title{Symmetric square root of a matrix}
\usage{
mhalf(M)
}
\arguments{
\item{M}{a positive semidefinite matrix}
}
\description{
Compute the symmetric square root of a matrix
}
\details{
This functions computes the symmetric square root
of a postive semidefinite matrix.
}
\examples{
S<-rwish(diag(5))
Sh<-mhalf(S)
Sh\%*\%Sh
}
\author{
Peter Hoff
}
\keyword{matrices}
