% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMLTR.R
\name{runMLTR}
\alias{runMLTR}
\title{Run Mcmc}
\usage{
runMLTR(y, x, NS = 7500, NB = 500, sdens = 100, plot = FALSE,
  seed = 6886, rstart = FALSE)
}
\arguments{
\item{Y}{some variable X some variable}
}
\value{
a list
}
\description{
More details
}
