% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfunctions.r
\name{tprod}
\alias{tprod}
\title{Tucker product}
\usage{
tprod(A, B, modes = 1:length(B))
}
\arguments{
\item{A}{a real valued array}

\item{B}{a list of matrices, the second dimension of each matching the
dimensions of A}

\item{modes}{a vector giving which modes of A should be multiplied}
}
\description{
Multiply an array by a list of matrices along each mode
}
\details{
This function multiplies an array along each mode.
}
\examples{
m<-c(6,5,4)
A<-rsan(c(6,5,4))
B<-list() ; for(k in 1:3) { B[[k]]<-rsan(c(3,dim(A)[[k]])) }
}
\author{
Peter Hoff
}
\keyword{arrays}
