% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfunctions.r
\name{tr}
\alias{tr}
\title{Matrix trace}
\usage{
tr(A)
}
\arguments{
\item{A}{a square matrix}
}
\description{
Compute the trace of a matrix
}
\details{
This function computes the trace of a square matrix.
}
\examples{
tr(riwsh(diag(5)))
}
\author{
Peter Hoff
}
\keyword{matrices}
