% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\name{yeast}
\alias{yeast}
\title{Yeast protien interactome}
\format{A binary, directed sociomatrix indicating which proteins bind to other
target protiens. Note that the diagonal is not missing.}
\source{
\url{http://interactome.dfci.harvard.edu/S_cerevisiae/download/Ito_core.txt}
}
\description{
Protein-protein interaction network in the yeast S. cerevisiae.
}
