% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfunctions.r
\name{zscores}
\alias{zscores}
\title{normal scores}
\usage{
zscores(y)
}
\arguments{
\item{y}{a vector}
}
\description{
Compute normal scores
}
\details{
This function applies a quantile-quantile transformation to
the data, resulting in a distribution that is approximately normal
but has the same ranks as the original data
}
\examples{
y<-rexp(100)
z<-zscores(y)
par(mfrow=c(1,3))
hist(y)
hist(z)
plot(y,z)
}
\author{
Peter Hoff
}
\keyword{transformation}
